moved RoommatesFilter into a folder
